Moved from bokeh to plotly
In previous implementations, the 2d visualizations were done using the bokeh package and
the 3d ones were done using matplotlib. Bokeh had the benefit of generating HTML files
which were self-contained and interactive. However, bokeh did not support 3d graphics
like matplotlib. Plotly proved to be the best of both packages as it supports 3d visuals
and can generate HTML files.

Furthermore, instead of the objective vector field (which was confusing when the axis
scales were different), there is now a togglable isoprofit line. This eliminates the
previous confusion with the vector field. NOTE: this has not been implemented in 3d.

Lastly, the ability to toggle between various pivot rules (in stand-alone HTML) was less
trivial in plotly. This functionality is no longer available in 2d as the bokeh 2d
implementation was replaced by plotly. This leaves a single simplex_visual function
which returns a plotly figure. It can be shown in a Jupyter notebook via .show() or an
HTML file can be created using .write_html(name).