Added style module
The style module was added which provides a higher level interface with
plotly. It includes functions for formatting numbers and equations, creating
labels, and plotting lines, ploygons, planes, and scatter plots. Furthermore,
it includes the function order which correctly orders a set of points for
plotting a 2d or 3d polygon.

Other misc. changes include:
- LP class now has get_basic_feasible_sol(B) for some potential basis B
- get_basic_feasible_solns now outputs the objective value of each point
- plot_constraint function removed (replaced with xy_from_equation)