Polished up style.py and added dictionary form tableau.
The transition of generic visual-related functions to style.py is now complete.
The module now has functions for properly formatting numbers and dictionaries
creating tables, setting axis limits, and plotting vectors, scatter plots,
lines, intersections, equations, and polygons. Furthermore, it contains the
function order which orders points for plotting purposes. Some of these
functions (such as polygon) take a style argument which specifies a
pre-defined style to use when plotting.

With this addition of the style module, separate plot_constraints and
plot_feasible_region functions became unnecessary and were removed. Their
function is now included within plot_lp.

The last significant change is the ability to choose between a tableau in
canonical or dictionary form. Dictionary form better translates to the
current ENGRI 1101 course material. The basic variables are all in the
LHS and the nonbasic variables as well as the constant are on the RHS.

Small misc. changes:
- A point is placed at the axes limits to fix axis rescaling bug
- The isoprofit plane now extends beyond the feasible region
- Each constraint in 3d is now plotted (muted by default)
- Arrow heads (previously invisible) removed