Intermediate iteration steps. Updated dictionary form tableau.
The differing tableau styles were removed from the simplex module. It
returned to previous implementation where get_tableau returns a numpy
array representing the canonical form tableau. There is now a function
in the visualize module which creates the header and content string
arrays to be passed to the table() function from the style module. This
function takes an argument specifying the form (canonical or dictionary)
of the desired tableau.

An intermediate step was added between simplex iterations. Geometrically,
the simplex path appears half-way between basic feasible solutions. In
addtion, both the tableau of the previous bfs and next bfs are shown.
This will hopefully allow students to better see the transition between
iterations of simplex.

Other misc. changes:
- The dictionary form better resembles current 1101 material
- New number and equation formatting method: linear_string()
- Fixed import statement bugs
- Added handful of global constants (height, width, etc..)