Finished linting.
All three modules have now been linted. Additionally, there is another
custom exception called InfiniteFeasibleRegion which is raised when an LP
is found to have an infinite feasible region and can not be visualized. The
docstring has been improved in various places and code in some places was
condensed and made generally clearer. NOTE: the vertical surface bug is
still an issue.