Added examples. More testing.
The previous test.py file was replaces with examples.py and numerous
examples have been created to emphasize different aspects of LPs in both
2d and 3d. Two of which are the Klee-Minty LP in 2d and 2d. Furthermore,
the package has been tested successfully on Python versions 3.5-3.8 using
the tox package. The tox.ini file is included in the MANIFEST.in file.
Lastly, the README has been updated to include badges connected to PyPI.