Added documentation.
Generated documentation using Sphinx. In addition to the documentation, the
site also includes installation instructions for general purpose and
development purposes, an introduction to linear programming and the simplex
method, a tutorial for showing usage, and a handful of examples. The
docstrings were adjusted to be more rst friendly. The majority of text on
the documentation is a first draft and has not been proofed.