Updated documentation.
The majority of this commit consists of changes to documentation. The
Cornell brand package was downloaded and applied to the Sphinx rtd theme.
The Quickstart Guide tutorial is now complete and includes many GILP
visualizations as examples. The main page of the docs were adjusted to
bring more attention to Cornell, the book, and the Plotly package. A custom
html layout was used to use Cornell red in the rtd theme. The example GILP
visuals were updated to reflect slight changes to code.

In addition to documentation changes, a few other misc. changes occured.
- An iteration counter was added to the canonical form LPs
- The iteration counter is highlighted red on the LPs
- The setup.py file was updated to require a higher version of Plotly
- Minor change to LICENSE file