Added lp_visual function.
The visualize module now has functions for creating an ispprofit slider
and an iteration slider. The functions are called isoprofit_slider and
iteration_slider and both return plt.layout.Slider objects. The
simplex_visual function will create a figure with both of these sliders
after running simplex. The lp_visual function creates a figure that
shows the geometry of an LP and only has the isoprofit slider.