Fixed student identified bugs.
Two bugs came up during testing. The first was that the origin was not
hoverable. This came from the fact that the shaded region layer was above
the BFS points. The BFS points are now the top layer. The other was that
the provided index options for the manual selection pivot rule were off by
one which is now fixed.