standard equality and inequality form changed.
LP class now has attribute equality which is true iff the LP is in standard
equality form and false if in inequality form. The equality form of an LP is
no longer stored within the LP class. Instead, a separate function called
equality_form returns an equivalent LP in standard equaluty form for some LP.

These changes were reflected everywhere in the rest of the program. Note that
simplex taking an initial feasible solution is now more complicated and will
require some reworking. Giving an initial feasible solution to simplex is
unstable currently.