Reworked initial_solution for simplex.
Previously, the inital basic feasible solution assumed the LP was in
standard inequality form. It only required the user to specify the value of
the decision variables (and not slack variables) when giving an intial
basic feasible solution. Now, that assumption is no longer made. Hence,
the full BFS must be given. The test cases were updated to reflect this. (in)
addition to some other changes that were not yet reflected in the tests.