Account for corner case in phase I.
After minimizing error in phase I of the simplex method, there may still be
basic artifical variables. Removing them goes one of two ways: either all
a_ij in the corresponding constraint are zero and the constraint can be
deleted or there is a nonzero a_ij and you must execute a pivot. The case
where the constraint is deleted is now handled and tested.