Changed intersection function.
The intersection function is now in the simplex module rather than the
style module. It no longer supports 2d intersections. Furthermore, an
interior point of the intersection may be passed to improve computation
time. This is used in creating the intersections between isoprofit planes
and the LP feasible region. The min and max value of the objective
function are found across the feasible region. All points achieving these
values are recorded and averaged to obtain a point s and t respectivley.
The line spanning s and t then forms a set of interior points which are
then passed to the intersection method to improve runtime. NOTE: if s and
t lie on the same edge, the line spanning between them will not consist of
interior points. This bug was identified but not yet addresed.