More program efficiencies.
Minimized number of calls to get_axis_limits (which is pretty slow) by
changing equation function to take domain as input rather than figure.
The domain represents the domain on which that equation should be displayed
and the limits of the plot are most often used as the domain. Furthermore,
more optimizations in setting plotly attributes were made. NOTE: I opened
an issue with Plotly to optimize their add_trace/add_traces function.