Start transition to pyhull qhull wrapper.
Since scipy does not return the facets by halfspaces like pyhull does,
the program is shifting to use pyull instead for getting halfspace
intersections. Currently, pyhull is used to find all basic feasible
solutions, find the feasible region of the 2d lp, and find the vertices
defining the facets of the feasible region for 3d lps. One of the
benefits of the switch is that pyhull returns 3d feasible region facet
vertices already sorted. Hence, a parameter was added to the polygon
function to indicate wheather or not the points passed are sorted.