Added halfspace_intersection function.
In generating the visualization of feasible solutions, the function
get_basic_feasible_solns is no longer called. Instead, a new function
halfspace_intersection is called. This function takes halfspaces in the
form of linear inequalities as inputs and computes their intersection. If
an interior point is not provided, one is generated. The pyhull
HalfspaceIntersection object has two attributes of significant importance:
the first gives all the vertices and the second gives the vertices defining
each facet of the intersection.