Added geometry module. Fixed interior point bug.
Previously, the interior point used to generate the isoprofit planes were
created by finding the (average) point at which the plane began and stopped
intersecting the feasible region s and t respectively. Then, the path
between s and t formed a set of interior points. However, if s and t were
on the same edge, the path did not contain any interior points. Now, the
points defining the intersection of the plane and feasible region are found
by adding the isoprofit plane as a halfspace and finding the halfspace
intersection. Since we now use pyhull, we can get the vertices with respect
to the isoprofit plane halfspace. Rather than generate a new interior point
for each isoprofit plane, the first interior point after intersecting the
feasible region can be used for all isoprofit planes that follow.

The geometry module was created with the following functions: intersection,
halfspace_intersection, interior_point, and order. This was done to remove
these functions from their previous modules where their functionality was
out of place. Furthermore, the code base was cleaned up (especially those)
functions in the geometry module) and linted.