Added test cases for geometry module.
Test cases for the interior_point and halfspace_intersection functions have
been written and the order test cases were moved from test_style to
test_geometry. In addition, the code coverage report was posted to codecov and
the codecov badge was added to the README. Lastly, some light linting was
done across the code base.