Fixed point order bug in plot_lp.
If the halfspace intersection had no interior point and the function
get_basic_feasible_solns was used to get the feasible region, pts could not
be assumed to be in order to form non-intersecting polygons. Hence, a variable
ordered was created to specify which of the two cases one was in. Furthermore,
a minor bug in the get_axis_limits function was fixed.