Fixed plot_lp feasible region section.
Similar to the previous bug, if halfspace intersection could not be used to
generate the feasible region, the ability to get facets_by_halfspace is not
there. Changed the order indicator to via_hs_intersection to identify what the
case was and re-implemented old code to draw planes from list of basic feasible
solutions. Also, fixed a test case for the interior point function.