Use trace_indices in visualize module.
Now that the Figure class exists, the visualize module is no longer responsible
for maintaing indices of traces of interest. Instead, the sets of traces can be
added under a chosen name. The get_indices function was added to Figure class.
It returns all indices under a given name, or (if specified) under all keys
that contain that name. Note, the ITERATION_STEP constant (which was 2) is now
permanatly 2 and not adjustable.