Split up plot_lp function.
The plot_lp function was split into add_feasible_region and add_constraints
functions which plot the feasible region and constraints of an LP respectivley.
The add_feasible_region function has the parameter set_axes which sets the axes
of the figure to accomodate the plotted feasible region if set to true.