Consolidate add_path, add_tableaus, and add_iter_slider
Created a single function called add_simplex_path which replaces the functions
add_path, add_tableaus, and add_iter_slider. It returns a plotly slider that
toggles between iterations of simplex. In addition to the figure and LP
parameters, it also takes parameters that can be passed to the simplex solver.