Adjusted return arguments of simplex function
Instead of returning the entire simplex path, now just returns the solution,
bases, value, and wheather or not it is optimal. This new implementation is
possible because of the add_simplex_path function which creates traces as it
goes through the iterations of simplex rather than relying on the simplex
function to return a whole path. The test cases were adjusted for this change.