Added ability to specify type of second subplot and bnb algo change.
The set_up_figure function now takes an additional argument type (with default
value of 'table') that indicates the type of the second subplot. If the second
subplot will display a branch and bound tree, the type should be 'scatter'.
Lastly, additional info about the constraints generated to branch on are
returned by the branch_and_bound_iteration function.