Plot tree (graph) on a Plotly figure.
The plot_tree function takes a networkX (which was added as a dependency) tree
and a specified root node and adds a visualization of the tree on to the
figure. The function contains the code to properly position the nodes of the
tree. Furthermore, attributes 'text' and 'color' can be specified on each node
and intepretted by the plotting function.