Arguments added to isoprofit and iteration slider functions.
The iteration function now takes an argument tableaus that is True when the
tableaus should be created for a simplex path and False otherwise. Both the
iteration and isoprofit sliders have a slider_pos argument which can take
values 'top' or 'bottom'. The isoprofit slider defaults to bottom while the
iteration slider defaults to top.