Improved positioning function within plot_network.
This positioing function aims to show the branches of the tree clearly while
still utilizing the small portion of the screen it has. If there are more than
4 nodes in the layer, the nodes are evenly spaces within the width. Otherwise,
they are put under their parent node and shifted a value proportional to their
depth in the tree. Lastly, the x values are perturbed to keep them properly
spaced and then shifted to ensure the x values lie within the width.