Moved (almost) all default plotly attributes to the visualize module.
This was part of an on-going effort to make the style module independent from
linear programming so it acts soley as higher level wrapper of the plotly
package. The trace creating functions table, scatter, line, equation, and
polygon take the following new form: In addition to the relevant parameters
for the function, they have a template argument which is a dictionary of
arguments for the trace. In addition, these functions can take any keyword
argument meant for the trace which will overide the value found in the template
argument if needed.

The visualize module now contains the constants for figure width, height, etc..
that were previously in the style module. Furthermore, there are constants for
every plotly trace template. These templates can are passed to the style module
trace creating functions to create traces in standard styles. Lastly, the
set_up_figure function now creates a new template rather than using the default
one. This was set up to minimize the number of arguments that needed to be
defined across the various plotly trace template constants.