BFS label aligned left and revised simplex used in interior_point.
The mode of linprog is interior-point by default. This was leading to some
numerical warnings so the mode was switched to revised simplex. Furthermore, a
tolerance of 1e-12 was built into the interior_point function where the radius
of the circumscribed circle must be greater than tol to be consifered feasible.