Show optimal solution on simplex path and cosmetic changes.
The optimal solution (filled in) and the initial solution (open) are both shown
always shown on the visualization. The add_feasible_region function takes an
argument called theme which can be light, dark, or outline. This is used to
create the LP feasible region outline and for highlighting the current node's
feasible region in the branch and bound visual.