Updated docstrings in style module and renamed to graphic.
Now that the style module is completely abstracted from linear programming, the
module just serves as a high-level interface with plotly. Hence, it was renamed
to graphic since it no longer contains the style-related constants that have
now been moved to the visualize module. Furthermore, small implementation
changes were made when possible to make code more concise and readable.