Bug fix: Round vertices and non-interior point intersection method.
In the case where an interior point can be used to compute the intersection,
the vertices are rounded to prevent numerical errors. This is now the case for
intersections when an interior point can not be used. Furthermore, in
generating the isoprofit planes, the case where an interior point can not be
used is now implemented. Note: an interior point can not be used when the
feasible region is lower dimensional.