Changed halfspace_intersection function. Pyhull dependency removed.
The halfspace_intersection function in the geometry module was altered so as
pyhull is no longer a dependency. The function uses the HalfspaceIntersection
class available in scipy.spatial instead and then computes the facets by
halfspace from the intersection vertices. The visualize module was adjusted
since this new implementation no longer garuntees ordered lists of points.