Added vertices and facets functions to the geometry module.
The vertices function computes the vertices of a given halfspace intersection.
An interior point can be provided if one is known to improve computation time.
The facets function returns a list of vertices defining each facet of a
halfspace intersection. The vertices can be provided if known to improve the
computation time.