Implemented add_feasible_region using new geometry and ploytope plot functions.
The add_feasible_region implementation is now cleaner and uses the
functionality of the polytope_vertices and polytope_facets functions from the
geometry module and the polytope function from the graphic module. Furthermore,
a seperate function was created for creating a scatter trace with hover labels
for every basic fesasible soultion.