Enhancement: no more to_equality_form(). Equality form stored in LP.
The to_equality_form function was removed in favor of having the LP class
maintain coefficents for the standard equality form LP. The function for
retrieving coefficents now takes an additonal parameter equality which
indicates whether standard equality or inequality form is returned. Lastly, the
n and m attributes denote the number of DVs and constraints.