Breaking up visualize module functions to enhance readability.
The functions used to add a feasible region and add constraints to a figure
have been replaced by ones that just generate the associated traces.
Furthermore, scale_axes was created to scale a figure's axes to fully view some
set of vertices and lp_vertices in the simplex module returns the vertices of
some inequality LP's feasible region. The lp_visual function was altered to
accommodate these changes. Furthermore, the simplex_visual function now calls
lp_visual and builds off of it to reduce redundancy in code.